feat: Added language

+++ app/services/storage.ts
@@ -1,38 +1,38 @@
 const storage = {
   set: (key: string, value: any) => {
+    if (typeof window === "undefined") return;
+
     if (!value || value.length <= 0) return;
 
-    if (key && value) {
-      if (typeof value === "string") {
-        return window.localStorage.setItem(key, value);
-      } else {
-        return window.localStorage.setItem(key, JSON.stringify(value));
-      }
+    if (typeof value === "string") {
+      return window.localStorage.setItem(key, value);
+    } else {
+      return window.localStorage.setItem(key, JSON.stringify(value));
     }
   },
-  get: (key: string) => {
-    if (key) {
-      return localStorage.getItem(key) as string;
-    }
-  },
-  remove: (key: string) => {
-    if (key) {
-      return window.localStorage.removeItem(key);
-    }
-  },
-  key: (index: number | string) => {
-    if (typeof index !== "number") {
-      return window.localStorage.key(Number(index));
-    }
 
-    if (typeof index === "number") {
-      return window.localStorage.key(index);
-    }
+  get: (key: string) => {
+    if (typeof window === "undefined") return null;
+    return window.localStorage.getItem(key) as string;
   },
+
+  remove: (key: string) => {
+    if (typeof window === "undefined") return;
+    return window.localStorage.removeItem(key);
+  },
+
+  key: (index: number | string) => {
+    if (typeof window === "undefined") return null;
+    return window.localStorage.key(Number(index));
+  },
+
   length: () => {
+    if (typeof window === "undefined") return 0;
     return window.localStorage.length;
   },
+
   clear: () => {
+    if (typeof window === "undefined") return;
     return window.localStorage.clear();
   },
 };
